+++ README.md
@@ -60,7 +60,7 @@
 
 ## Application
 - [x] 애플리케이션을 재시작하더라도 이전에 하던 체스 게임을 다시 시작할 수 있어야 한다.
- - [ ] 종료된 게임을 불러오면 종료되었다고 알린다.
+ - [x] 종료된 게임을 불러오면 종료되었다고 알린다.
 - [x] DB를 적용할 때 도메인 객체의 변경을 최소화해야한다.
 - [x] 체스 게임방을 만들고 체스 게임방에 입장할 수 있는 기능을 추가한다.
   - [x] 사용자를 생성할 수 있다.
