+++ src/main/java/chess/Application.java
@@ -3,8 +3,13 @@ package chess;
 import chess.controller.GameController;
 import chess.controller.RoomController;
 import chess.controller.UserController;
-import chess.infra.JdbcConnectionPool;
-import chess.repository.*;
+import chess.database.JdbcConnectionPool;
+import chess.repository.MovementDao;
+import chess.repository.MovementRepository;
+import chess.repository.RoomDao;
+import chess.repository.RoomRepository;
+import chess.repository.UserDao;
+import chess.repository.UserRepository;
 import chess.service.GameService;
 import chess.service.RoomService;
 import chess.service.UserService;
