+++ src/main/java/chess/database/DatabaseConfiguration.java
@@ -1,4 +1,4 @@
-package chess.infra;
+package chess.database;
 
 public class DatabaseConfiguration {
