+++ src/main/java/chess/database/JdbcConnectionPool.java
@@ -1,4 +1,4 @@
-package chess.infra;
+package chess.database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
