+++ src/main/java/chess/repository/MovementDao.java
@@ -1,9 +1,12 @@
 package chess.repository;
 
+import chess.database.JdbcConnectionPool;
 import chess.domain.square.Movement;
-import chess.infra.JdbcConnectionPool;
-
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
