+++ src/main/java/chess/repository/UserDao.java
@@ -1,9 +1,12 @@
 package chess.repository;
 
+import chess.database.JdbcConnectionPool;
 import chess.domain.user.User;
-import chess.infra.JdbcConnectionPool;
-
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
